requirements.txt
python-dotenv
fastapi
uvicorn[standard]
numpy
scipy
tensorflow
librosa
h5py
python-multipart
